+++ src/components/SideBar/SideBar.jsx
@@ -1,44 +1,21 @@
 import { useQuery } from '@tanstack/react-query';
 import { useCallback, useEffect, useState } from 'react';
-import Detail from '../../pages/DetailPage/Detail';
-import {
-  InputAderss,
-  InputName,
-  ModalContent,
-  ModalOverlay,
-  SideBarContainer,
-  SideBarItem,
-  SideBarMenu,
-  SideBarMenuItem
-} from '../../styles/SideBar/sideBarStyle';
-import { SearchCloseButton, SideBarButton, SideBarDetailBtn } from '../../styles/common/btnStyle';
+import { SideBarContainer, SideBarMenu } from '../../styles/SideBar/sideBarStyle';
+import { SideBarButton } from '../../styles/common/btnStyle';
 import supabase from '../../supabase/supabaseClient';
-import DetailCarousel from '../DetailCarousel';
 import Pagination from './Pagination';
 import Search from './Search';
+import Item from '../Item';
 
 const SideBar = ({ setFilteredShops, setSelectedShop }) => {
   const [isOpen, setIsOpen] = useState(true);
   const [filteredShops, setFilteredShopsLocal] = useState([]);
-  const [modalStates, setModalStates] = useState({});
   const [page, setPage] = useState(1);
   const [totalPages, setTotalPages] = useState(1);
   const [currentShops, setCurrentShops] = useState([]);
 
   const toggleSidebar = () => {
     setIsOpen(!isOpen);
-  };
-
-  const toggleModal = (shopId) => {
-    setModalStates((prevStates) => ({
-      ...prevStates,
-      [shopId]: !prevStates[shopId]
-    }));
-  };
-
-  const stopBubble = (e) => {
-    e.stopPropagation();
-    e.preventDefault();
   };
 
   const fetchRestaurants = useCallback(async () => {
@@ -75,8 +52,7 @@ const SideBar = ({ setFilteredShops, setSelectedShop }) => {
   }, [page, filteredShops]);
 
   if (isPending) {
-    <div>loading..</div>;
-    return;
+    return <div>loading..</div>;
   }
 
   if (isError) {
@@ -97,32 +73,7 @@ const SideBar = ({ setFilteredShops, setSelectedShop }) => {
       />
       <SideBarMenu>
         {currentShops.length > 0 ? (
-          currentShops.map((shop) => (
-            <SideBarMenuItem key={shop.id} onClick={() => setSelectedShop(shop)}>
-              <SideBarItem>
-                <InputName>{shop.name}</InputName>
-                <InputAderss>
-                  <label htmlFor="adress">주소: </label>
-                  {shop.address}
-                </InputAderss>
-                <ul>{shop.loaction}</ul>
-              </SideBarItem>
-              <DetailCarousel shop={shop} />
-              <SideBarDetailBtn type="button" onClick={() => toggleModal(shop.id)}>
-                상세보기
-              </SideBarDetailBtn>
-              {modalStates[shop.id] && (
-                <ModalOverlay onClick={() => toggleModal(shop.id)}>
-                  <ModalContent onClick={stopBubble}>
-                    <SearchCloseButton type="button" onClick={() => toggleModal(shop.id)}>
-                      X
-                    </SearchCloseButton>
-                    <Detail shop={shop} />
-                  </ModalContent>
-                </ModalOverlay>
-              )}
-            </SideBarMenuItem>
-          ))
+          currentShops.map((shop) => <Item key={shop.id} shop={shop} />)
         ) : (
           <p>검색 결과가 없습니다.</p>
         )}
