+++ src/pages/MainPage/MainPage.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import Map from '../../components/Map';
 import SideBar from '../../components/SideBar';
-import { StMain } from '../../styles/mainPage/mainPageStyle';
+import { StMain } from '../../styles/MainPage/mainPageStyle';
 
 const MainPage = () => {
   const [filteredShops, setFilteredShops] = useState([]);
