+++ src/styles/SideBar/sideBarStyle.js
@@ -77,6 +77,7 @@ export const ModalOverlay = styled.div`
   display: flex;
   justify-content: center;
   align-items: center;
+  z-index: 1000;
 `;
 
 export const ModalContent = styled.div`
@@ -89,6 +90,7 @@ export const ModalContent = styled.div`
   max-width: 90%;
   overflow-y: auto;
   max-height: 90%;
+  z-index: 1001;
 
   &::-webkit-scrollbar {
     width: 12px;
